Changed percent to int from short

+++ event_driven_sim/stats_structs.h
@@ -6,7 +6,7 @@
 #define TIME_TABLE_SIZE 1920
 
 typedef int Customers;
-typedef short Percent;
+typedef int Percent;
 typedef int Time;
 
 typedef struct ArrivalData{
